requirements.txt
Django==5.2.3
pillow==11.2.1
django-admin-sortable2==2.2.8
django-tinymce==4.1.0
environs==14.2.0
requests==2.32.4